Conflicts:
	GitX.xcodeproj/project.pbxproj

+++ build_libgit2.sh
@@ -12,13 +12,16 @@ buildAction () {
     echo "Building libgit2..."
 	if [[ -d .git ]]
 	then
-		export PATH=$PATH:$HOME/bin:$HOME/local/bin:/sw/bin:/opt/local/bin:`"$TARGET_BUILD_DIR"/gitx --git-path`
+        if [ "$(which git)" == "" ]; then
+            echo "git not found"
+            exit -1
+        fi
 		git submodule init
 		git submodule sync
 		git submodule update
 		cd libgit2
 		rm -f libgit2.a
-		make CFLAGS="-arch i386 -arch x86_64"
+		make CFLAGS="-arch i386 -arch ppc"
 		ranlib libgit2.a
 	else
 		echo "error: Not a git repository."
